+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Update module structure :
+  - Move Controllers/ to Http/
+- Update MakeModuleControllerCommand
+- Update ModuleCreateCommand
+
 ## [1.0.0] - 2025-10-05
 
 ### Added